+++ spec/dummy/src/models/account.js
@@ -3,6 +3,8 @@ import AccountBase from "../model-bases/account.js"
 class Account extends AccountBase {
 }
 
-Account.setDatabaseIdentifier("mssql")
+if (process.env.VELOCIOUS_DISABLE_MSSQL !== "1") {
+  Account.setDatabaseIdentifier("mssql")
+}
 
 export default Account