+++ src/environment-handlers/browser.js
@@ -129,6 +129,7 @@ export default class VelociousEnvironmentsHandlerBrowser extends Base {
     let filePath = "."
 
     for (let commandPart of commandParts) {
+      if (commandPart == "c") commandPart = "console"
       if (commandPart == "d") commandPart = "destroy"
       if (commandPart == "g") commandPart = "generate"
       if (commandPart == "s") commandPart = "server"